metrics_ui/api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
